+++ src/modelo/Clientes.java
@@ -1,12 +1,13 @@
 package modelo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Clientes implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 	String nomeCliente;
-	int cpf;
+	BigDecimal cpf;
 	
 	public String getNomeCliente() {
 		return nomeCliente;
@@ -16,11 +17,11 @@ public class Clientes implements Serializable {
 		this.nomeCliente = nome;
 	}
 
-	public int getCpf() {
+	public BigDecimal getCpf() {
 		return cpf;
 	}
 
-	public void setCpf(int cpf) {
+	public void setCpf(BigDecimal cpf) {
 		this.cpf = cpf;
 	}
